Testing search method
